requirements.txt
black
dash
dash-bootstrap-components
dash-multi
chart_studio
Flask-Caching
flake8
gunicorn
jitcache
pandas
psutil
scikit-learn
xlrd
requests
inflect
matplotlib